CLI/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autosign-buaa-cli"
version = "0.1.0"
description = "BUAA iClass AutoSign command-line tool"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "DeNeRATe-cool" }
]
keywords = ["buaa", "iclass", "autosign", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Operating System :: OS Independent",
  "Environment :: Console",
]
dependencies = [
  "requests>=2.31,<3.0",
  "beautifulsoup4>=4.12,<5.0",
  "PyYAML>=6.0,<7.0",
]

[project.urls]
Homepage = "https://github.com/DeNeRATe-cool/AutoSign"
Repository = "https://github.com/DeNeRATe-cool/AutoSign"

[project.scripts]
autosign = "autosign_cli.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
